refactor(google): resource migrations batch 11
Migrate sql_database_instance alongside few case fix in registry file.
